+++ src/main/java/com/rifatul/SpringBucks/exception/ProductAlreadyOnCartException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
 
 public class ProductAlreadyOnCartException extends ResponseStatusException {
-    public ProductAlreadyOnCartException(int orderId, int productId) {
+    public ProductAlreadyOnCartException(long orderId, int productId) {
         super(HttpStatus.CONFLICT, "the cart #" + orderId + " is already contains the product #" + productId);
     }
 }
